+++ xmake/rules/utils/bin2c/xmake.lua
@@ -20,12 +20,18 @@
 
 rule("utils.bin2c")
     set_extensions(".bin")
+    on_load(function (target)
+        local headerdir = path.join(target:autogendir(), "rules", "c++", "bin2c")
+        if not os.isdir(headerdir) then
+            os.mkdir(headerdir)
+        end
+        target:add("includedirs", headerdir)
+    end)
     before_buildcmd_file(function (target, batchcmds, sourcefile_bin, opt)
 
         -- get header file
         local headerdir = path.join(target:autogendir(), "rules", "c++", "bin2c")
         local headerfile = path.join(headerdir, path.filename(sourcefile_bin) .. ".h")
-        target:add("includedirs", headerdir)
 
         -- add commands
         batchcmds:show_progress(opt.progress, "${color.build.object}generating.bin2c %s", sourcefile_bin)